Add new federation flag to community structure

+++ include/n2n.h
@@ -359,7 +359,7 @@ typedef struct sn_stats
 struct sn_community
 {
   char community[N2N_COMMUNITY_SIZE];
-  int  is_federation;     		      /* if not-zero, then the current community is the federation of supernodes */
+  uint8_t             is_federation;          /* if not-zero, then the current community is the federation of supernodes */
   uint8_t             purgeable;              /* indicates purgeable community (fixed-name, predetermined (-c parameter) communties usually are unpurgeable) */
   uint8_t	      header_encryption;      /* Header encryption indicator. */
   he_context_t        *header_encryption_ctx; /* Header encryption cipher context. */
